Cloak of Sound Rest fixes
Basically, the cloak was set up to block any attack by maing the wearer immune to magical and non-magical weapons, which would be cleared
by a triggered on-hit op 232. However, blocking weapons via op 120 means that they technically never hit, so the wearer was effectively
immune to all weapons. EEex does have a way to allow these to 'hit' despite the protections, however, even with EEex there are loopholes
that would allow the protection to persist due to how op 232 works.

I played around with other schemes--the cloak providing a single stoneskin, or 100% physical damage resistance--but each one either
didn't interact well with op 232 or had the same loophole that allowed the protection to persist through multiple hits. Instead I've
reworked the cloak to provide regeneration if struck and updated the description.